v3.6.0-rc.1: split tools.ts (4252 lines) into src/tools/ domain modules
v3.6.0 Phase 1 of 4. Pure refactor: same exported surface, same
signatures, all 712 tests pass. Closes the #1 finding across 3
independent external audits (Mavis ×2 + MiniMax) — the tools.ts
monolith was flagged in every audit as the only blocker between
the project and a clean top-tier rating.

Structure:
- src/tools/search.ts  (1224 lines, 19 exports)  — 4 search variants
                                                   + TF-IDF helpers
- src/tools/write.ts   ( 682 lines, 21 exports)  — 6 write tools
                                                   + rename helpers
- src/tools/read.ts    ( 864 lines, 28 exports)  — read/list/links
                                                   + frontmatter + chat
- src/tools/media.ts   ( 516 lines, 16 exports)  — pdf + canvas + ocr
- src/tools/meta.ts    ( 984 lines, 26 exports)  — contextPack/validate
                                                   + lint + path
- src/tools/index.ts   (   5 lines, barrel)      — re-exports all

Helper visibility (8 helpers previously private, now exported for
cross-domain use — all benign cycles, only called inside async fn
bodies):
- meta: jaccard, intersectionSize, ngrams, indexFor, findBestMatch,
        stripMd, normalizeTag
- search: sliceSnippet, tokenizeForTfidf, buildTfidfIndex
- write: composeNote, extractFrontmatterTagsLower,
         resolvePeriodicAlias, suggestSimilar, resolveTarget,
         rewriteRawTarget, rewriteOutsideCodeFences,
         replaceStringOutsideCodeFences

None of these are in package.json `exports` map, so no SemVer
contract change. STABILITY.md surface unaffected.

Import path updates:
- src/index.ts:   "./tools.js" → "./tools/index.js"
- src/eval.ts:    "./tools.js" → "./tools/index.js"
- 15 test files:  same bulk update

Also: CLAUDE.md added at repo root — fixes the v3.6.0 sprint goal,
scope, quality bar, exit criteria, non-goals + anti-patterns so any
session on this repo shares the same North Star.

712 tests pass · lint clean · tsc strict clean · smoke pass ·
version-consistency green at 3.6.0-rc.1 (5 surfaces).

npm dist-tag: rc (NOT latest). Users on latest remain on v3.5.14.
Install: npm i @oomkapwn/enquire-mcp@rc.

Next: v3.6.0-rc.2 — split index.ts (3665 lines).
