+++ src/index.ts
@@ -49,11 +49,11 @@ import {
   searchText,
   semanticSearch,
   validateNoteProposal
-} from "./tools.js";
+} from "./tools/index.js";
 import { Vault } from "./vault.js";
 import { VaultWatcher } from "./watcher.js";
 
-const VERSION = "3.5.14";
+const VERSION = "3.6.0-rc.1";
 
 /** Default location for the persistent embedding index, alongside .fts5.db. */
 function embedDbPath(vaultRoot: string): string {
