+++ tests/chat-thread.test.ts
@@ -6,7 +6,7 @@ import { promises as fs } from "node:fs";
 import * as os from "node:os";
 import * as path from "node:path";
 import { afterEach, beforeEach, describe, expect, it } from "vitest";
-import { chatThreadAppend, chatThreadRead } from "../src/tools.js";
+import { chatThreadAppend, chatThreadRead } from "../src/tools/index.js";
 import { Vault } from "../src/vault.js";
 
 let root: string;
