+++ tests/ocr.test.ts
@@ -18,7 +18,7 @@ import * as os from "node:os";
 import * as path from "node:path";
 import { afterEach, beforeEach, describe, expect, it } from "vitest";
 import { extractPdfWithOcr, isOcrAvailable } from "../src/ocr.js";
-import { ocrPdf } from "../src/tools.js";
+import { ocrPdf } from "../src/tools/index.js";
 import { Vault } from "../src/vault.js";
 import { makePdf } from "./helpers/make-pdf.js";
