+++ tests/pdf.test.ts
@@ -15,7 +15,7 @@ import * as os from "node:os";
 import * as path from "node:path";
 import { afterEach, beforeEach, describe, expect, it } from "vitest";
 import { extractPdfText, isPdfjsAvailable } from "../src/pdf.js";
-import { listPdfs, readPdf } from "../src/tools.js";
+import { listPdfs, readPdf } from "../src/tools/index.js";
 import { Vault } from "../src/vault.js";
 import { makePdf } from "./helpers/make-pdf.js";
