+++ tests/periodic.test.ts
@@ -3,7 +3,7 @@ import * as os from "node:os";
 import * as path from "node:path";
 import { afterEach, beforeEach, describe, expect, it } from "vitest";
 import { formatMoment, loadPeriodicConfig, resolvePeriodicNoteName } from "../src/periodic.js";
-import { readNote } from "../src/tools.js";
+import { readNote } from "../src/tools/index.js";
 import { Vault } from "../src/vault.js";
 
 let root: string;
