+++ tests/search-hybrid.test.ts
@@ -9,7 +9,7 @@ import * as os from "node:os";
 import * as path from "node:path";
 import { afterAll, beforeAll, describe, expect, it } from "vitest";
 import { defaultIndexFile, FtsIndex } from "../src/fts5.js";
-import { searchHybrid } from "../src/tools.js";
+import { searchHybrid } from "../src/tools/index.js";
 import { Vault } from "../src/vault.js";
 
 let root: string;
