+++ tests/write.test.ts
@@ -2,7 +2,7 @@ import { promises as fs } from "node:fs";
 import * as os from "node:os";
 import * as path from "node:path";
 import { afterEach, beforeEach, describe, expect, it } from "vitest";
-import { appendToNote, archiveNote, createNote, renameNote, replaceInNotes } from "../src/tools.js";
+import { appendToNote, archiveNote, createNote, renameNote, replaceInNotes } from "../src/tools/index.js";
 import { Vault } from "../src/vault.js";
 
 let root: string;
